Make OPEN, LOAD and SAVE aware of directories
The GW-BASIC manual documents the VARPTR function as returning a
pointer to an FCB. The observed behavior of GW-BASIC 3.23 is that
this pointer points to a proper FCB.

To match this behavior, the FCB-based functions are kept. If a
directory is requested, we temporarily change to that directory
for the duration of the statement.